Consolidate Containers
No verify
